Refactor monorepo and demo layout
