+++ index.html
@@ -15,7 +15,7 @@
     <meta name="msapplication-TileColor" content="#da532c" />
     <meta name="theme-color" content="#ffffff" />
 
-    <link rel="stylesheet" href="./src/styles.scss" />
+    <link rel="stylesheet" href="./src/_lib/styles.scss" />
     <style>
       :root {
         --bg: #222;
@@ -49,13 +49,13 @@
   <body>
     <ul class="index">
       <li class="sample contour">
-        <a href="/examples/contour.html"><span>Contour</span></a>
+        <a href="./src/contour/index.html"><span>Contour</span></a>
       </li>
       <li class="sample seigaiha">
-        <a href="/examples/seigaiha.html"><span>Seigaiha</span></a>
+        <a href="./src/seigaiha/index.html"><span>Seigaiha</span></a>
       </li>
       <li class="sample truchet">
-        <a href="/examples/truchet.html"><span>Truchet</span></a>
+        <a href="./src/truchet/index.html"><span>Truchet</span></a>
       </li>
     </ul>
