+++ src/truchet/vite.config.js
@@ -1,3 +1,3 @@
-import { getConfig } from "../vite.config.base.js";
+import { getConfig } from "../_lib/vite.config.base.js";
 
 export default getConfig(import.meta.url);
