+++ tsconfig.json
@@ -7,6 +7,6 @@
     "checkJs": false,
     "noEmit": true
   },
-  "include": ["typings/*", "src/**/*", "lib/utils.ts"],
+  "include": ["typings/*", "src/**/*"],
   "exclude": [".eslintrc.cjs"]
 }
